Adds a couple tests demonstrating how apps that must execute synchronously would call into this library, including a simulation of how the main message pump works.
The synchronous test would hang without the .ConfigureAwait(false) changes made in the prior commit, if WinRT happens to decide to yield in their async method, which I've found happens intermittently.

A yield could be forced in one of these async methods of the library by adding this line:
await Task.Delay(0);
Adding this line to one of the methods called by this test would cause the test to hang. Modifying this line to instead read:
await Task.Delay(0).ConfigureAwait(false);
resolves the hang, demonstrating what the effect of ConfigureAwait(false) does.

Note however that every line in the test code actually runs on the main thread. The only code that might run on a threadpool thread is the code that is internal to this library, which has no thread-safety requirements anyway (at least none that were fulfilled before this change).