Adds CancellationToken parameter to all async methods.
This allows async methods in app code that have a cancellation token to propagate that cancellation token to the I/O code. This is important since I/O code *can* be quite slow, and if a Cancel button is exposed to the user it should be responsive when the user taps it and cancel right away rather than wait for the I/O operation to finish.

For those platforms that natively support CancellationToken in their I/O stack, we simply propagate it to them. Otherwise, we check the token first and then proceed with the non-cancellable I/O operation.

This new parameter is optional, so there is no source code breaking change here, but there is a binary breaking change since a recompile is necessary.